requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
numpy
xgboost
shap
scikit-learn
requests
streamlit
plotly
